change name of package canutils to can-utils to get recent versions
The package name "canutils" refers to an older/alternative version of can-utils with a different command line interface than recent versions you get from package management of popular linux distros. Besides that this old version does not support CAN-FD, so you can not dump or send CAN-FD messages, even when the hardware supports this. This little change is enough, because the newer "can-utils" GitHub repo (https://github.com/linux-can/can-utils) is already in the Yocto sources.